requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langchain-google-genai==2.0.8
sentence-transformers==3.3.1
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.41.1
faiss-cpu==1.9.0.post1